Move `MetaSuggestionsQueryHandler` to `stroom-core` module
